Move version from wagtail.wagtailcore to wagtail
This is where developers expect it to be, similar to Django and other
projects. The version info still exists at the old `wagtail.wagtailcore`
location, for backwards compatibility.

Fixes #2557
