+++ wagtail/wagtailcore/__init__.py
@@ -1,15 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
-from wagtail.utils.version import get_semver_version, get_version
-
-# major.minor.patch.release.number
-# release must be one of alpha, beta, rc, or final
-VERSION = (1, 7, 0, 'alpha', 0)
-
-__version__ = get_version(VERSION)
-
-# Required for npm package for frontend
-__semver__ = get_semver_version(VERSION)
+# Imported for historical reasons
+from wagtail import __semver__, __version__  # noqa
 
 
 def setup():
